requirements.txt
streamlit==1.38.0
tensorflow==2.17.0
numpy==1.26.4
pillow==10.4.0
